requirements.txt
# =============================================================================
# TradingAgents - Pinned Dependencies
# =============================================================================
# All versions are pinned to prevent supply chain attacks and breaking changes
# Last updated: 2025-11-17
# =============================================================================

# Core LangChain Framework
# -----------------------------------------------------------------------------
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-experimental
langgraph

# AI/ML API Clients
# -----------------------------------------------------------------------------
openai

# Data & Financial APIs
# -----------------------------------------------------------------------------
pandas
numpy
yfinance
eodhd
akshare
tushare
finnhub-python

# Web Scraping & Parsing
# -----------------------------------------------------------------------------
requests
feedparser
parsel
praw
gnews

# Technical Analysis & Backtesting
# -----------------------------------------------------------------------------
stockstats
backtrader
scikit-learn

# Database & Caching
# -----------------------------------------------------------------------------
chromadb
redis
pyarrow  # Required for secure Parquet-based caching

# Web Interface
# -----------------------------------------------------------------------------
chainlit

# CLI & UI Utilities
# -----------------------------------------------------------------------------
rich
questionary
tqdm

# System Utilities
# -----------------------------------------------------------------------------
python-dotenv
pytz
typing-extensions
setuptools

# Testing & Development
# -----------------------------------------------------------------------------
pytest
pytest-cov
pytest-asyncio
google-generativeai
